requirement.txt
streamlit
pdfplumber
python-docx
langchain
keybert
sentence-transformers
spacy
nltk
requests
python-dotenv